fix: Start refactoring and adding the deployment scripts from -siwe version
