+++ .env.example
@@ -1,0 +1,7 @@
+NETWORK=ic
+IDENTITY=default
+SALT=
+CHAIN_ID=
+STATEMENT=
+URI=
+DOMAIN=
