+++ .env.local.example
@@ -1,0 +1,7 @@
+NETWORK=local
+IDENTITY=default
+SALT=verysecret1234
+CHAIN_ID=8453
+STATEMENT=Login to localhost
+URI=http://localhost:3000
+DOMAIN=localhost
