+++ Cargo.toml
@@ -21,7 +21,7 @@ opt-level = "s"
 
 [workspace.dependencies]
 hex = { version = "0.4.3", features = ["serde"] }
-ic_bls12_381 = { version = "0.10.0", default-features = false, features = [
+ic_bls12_381 = { version = "0.10.1", default-features = false, features = [
     "groups",
     "pairings",
     "alloc",
