--- canister_ids.json
@@ -1,11 +1,0 @@
-{
-  "vetkd_notes": {
-    "ic": "nr2en-niaaa-aaaao-a3tsq-cai"
-  },
-  "vetkd_system_api": {
-    "ic": "nn664-2iaaa-aaaao-a3tqq-cai"
-  },
-  "vetkd_www": {
-    "ic": "nd4tu-byaaa-aaaao-a3trq-cai"
-  }
-}
