+++ deploy.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+pnpm deploy:ic -- "${@}"
