+++ packages/vetkd-notes/Cargo.toml
@@ -14,11 +14,11 @@ wasm-opt = false
 candid = "0.10.13"
 ic-cdk = "0.17.1"
 ic-cdk-macros = "0.17.1"
-ic-stable-structures = "0.6.7"
+ic-stable-structures = "0.6.8"
 lazy_static = "1.5.0"
-serde_json = "1.0.138"
-anyhow = "1.0.95"
-serde = "1.0.217"
+serde_json = "1.0.140"
+anyhow = "1.0.97"
+serde = "1.0.219"
 ic-types = "0.7.0"
 hex = "0.4.3"
 custom_debug = "0.6.2"
